extracting tasks and commandline arguments
